@@ -1,51 +1,148 @@
+plugins {
+    id "de.develappers.versioning" version "1.0.16"
+}
+
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
+apply plugin: 'io.fabric'
+apply plugin: 'realm-android'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "de.htwdd.htwdresden"
-        minSdkVersion 21
-        targetSdkVersion 27
-        versionCode 20
-        versionName "1.5.1"
-        testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
+        minSdkVersion 23
+        targetSdkVersion 29
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
-    buildTypes {
+
+    def keystorePropertiesFile = rootProject.file("keystore.properties")
+    def keystoreProperties = new Properties()
+    keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
+
+    signingConfigs {
         release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            storeFile file(keystoreProperties['file'])
+            storePassword keystoreProperties['password']
+            keyAlias keystoreProperties['alias']
+            keyPassword keystoreProperties['password']
         }
     }
+
+    buildTypes {
+        release {
+            debuggable false
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            signingConfig signingConfigs.release
+        }
+
+        debug {
+            debuggable true
+        }
+    }
+
     lintOptions {
         disable 'MissingTranslation'
+        checkReleaseBuilds false
+        abortOnError false
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+
+    dataBinding {
+        enabled = true
+    }
+}
+
+realm {
+    kotlinExtensionsEnabled = true
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support.constraint:constraint-layout:1.1.0'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
-    // https://github.com/ACRA/acra
-    implementation "ch.acra:acra-mail:5.1.3"
-    implementation "ch.acra:acra-notification:5.1.3"
-    implementation "ch.acra:acra-dialog:5.1.3"
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    //noinspection GradleCompatible
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+
     // https://github.com/realm/realm-android-adapters
     implementation 'io.realm:android-adapters:2.1.1'
+
     // https://github.com/square/retrofit
-    implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
+    def retrofit2_version = "2.6.1"
+    implementation "com.squareup.retrofit2:converter-gson:$retrofit2_version"
+    implementation "com.squareup.retrofit2:retrofit:$retrofit2_version"
+    implementation "com.squareup.retrofit2:adapter-rxjava2:$retrofit2_version"
+
     // https://square.github.io/picasso/
     implementation 'com.squareup.picasso:picasso:2.71828'
+
     // https://github.com/heinrichreimer/android-canteen-balance
     implementation 'com.heinrichreimersoftware:android-canteen-balance:0.1.2'
-    apply plugin: 'realm-android'
+
+    implementation 'com.github.msayan:tutorial-view:v1.0.10'
+
+    implementation 'com.google.firebase:firebase-core:17.2.1'
+    implementation 'com.crashlytics.sdk.android:crashlytics:2.10.1'
+
+    //Lottie
+    implementation 'com.airbnb.android:lottie:3.0.1'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    // Android Kotlin Core Extensions
+    implementation 'androidx.core:core-ktx:1.1.0'
+
+    // Android Security
+    def security_version = "1.0.0-alpha02"
+    implementation "androidx.security:security-crypto:$security_version"
+
+    // Android Lifecycle
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.1.0'
+
+    // Android RX
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.10'
+    implementation 'io.reactivex.rxjava2:rxkotlin:2.4.0'
+
+    // Navigation Components
+    def nav_version = "2.2.0-alpha02"
+    implementation "androidx.navigation:navigation-fragment:$nav_version"
+    implementation "androidx.navigation:navigation-ui:$nav_version"
+    implementation "androidx.navigation:navigation-fragment-ktx:$nav_version"
+    implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
+
+    // Material Dialogs
+    def md_version = "3.1.0"
+    implementation "com.afollestad.material-dialogs:core:$md_version"
+    implementation "com.afollestad.material-dialogs:input:$md_version"
+    implementation "com.afollestad.material-dialogs:bottomsheets:$md_version"
+
+    // Toasts with steroids
+    def toast_version = "1.4.2"
+    implementation "com.github.GrenderG:Toasty:$toast_version"
+
+    // Konfetti
+    def konfetti_verson = "1.1.3"
+    implementation "nl.dionsegijn:konfetti:$konfetti_verson"
+
+    // Zoomable ImageView
+    implementation 'com.jsibbold:zoomage:1.3.0'
+}
+apply plugin: 'com.google.gms.google-services'
+repositories {
+    mavenCentral()
 }
