@@ -1,10 +1,11 @@
 package de.htwdd.htwdresden.adapter;
 
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.Nullable;
 
 import de.htwdd.htwdresden.interfaces.ISpinnerEntity;
 import io.realm.OrderedRealmCollection;
