@@ -9,6 +9,7 @@ public final class Const {
 
     public static final class BundleParams {
         public static final String MENSA_DETAIL_MODE = "MENSA_DETAIL_MODE";
+        public static final String CANTEEN_DETAIL_MODE = "CANTEEN_DETAIL_MODE";
         public static final String TIMETABLE_WEEK = "TIMETABLE_WEEK";
         public static final String TIMETABLE_DAY = "TIMETABLE_DAY";
         public static final String TIMETABLE_DS = "TIMETABLE_DS";
@@ -34,6 +35,8 @@ public final class Const {
         static final String PREFERENCES_AUTO_MUTE_MODE = "autoMuteMode";
         public static final String PREFERENCES_SEMESTERPLAN_UPDATETIME = "semesterPlanUpdateTime";
         public static final String PREFERENCES_MENSA_WEEK_LASTUPDATE = "mensaWeekLastUpdate";
+        public static final String PREFERENCES_MENSA_NEXT_WEEK_LASTUPDATE = "mensaWeekLastUpdate";
+        public static final String PREFERENCES_MENSA_DAY_LASTUPDATE = "mensaDayLastUpdate";
         public static final String PREFERENCES_STUDY_GROUP_LAST_UPDATE = "studyGroupsLastUpdate";
         public static final String PREFERENCES_AUTO_EXAM_UPDATE = "autoExamUpdate";
         public static final String PREFERENCES_TIMETABLE_STUDIENJAHR = "studyGroupYear";
@@ -97,7 +100,7 @@ public final class Const {
             public static final String MENSA_ID = "mensaId";
             public static final String MENSA_DATE = "date";
             public static final String MENSA_IS_SOLDOUT = "isSoldOut";
-            public static final String MENSA_IMAGE = "image";
+            public static final String MENSA_CATEGORY = "category";
         }
 
         public static class ExamResults {
