@@ -2,10 +2,11 @@ package de.htwdd.htwdresden.types;
 
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 /**
  * Die Klasse beschreibt (charakterisiert) einen Tab im ViewPagerAdapter
