@@ -1,16 +1,20 @@
 package de.htwdd.htwdresden.types.canteen;
 
+import io.realm.RealmList;
+import io.realm.RealmObject;
+
 /**
  * Beschreibung einer Kantine
  *
  * @author Kay Förster
  */
-public class Canteen {
+public class Canteen extends RealmObject {
     private int id;
     private String city;
     private String name;
     private String address;
-    private Coordinates coordinates;
+    private RealmList<Float> coordinates;
+    private Boolean isFav;
 
     public int getId() {
         return id;
@@ -28,7 +32,14 @@ public class Canteen {
         return address;
     }
 
-    public Coordinates getCoordinates() {
+    public RealmList<Float> getCoordinates() {
         return coordinates;
+    }
+
+    public Boolean getFav() {
+        return isFav;
+    }
+    public void setIsFav(Boolean isFav) {
+        this.isFav = isFav;
     }
 }
